Add method to enable setting ptr on VirConnection
Since ptr is not exported, we need a function to allow it to be set
by another package.

Also, since the type of ptr is derived from C, other packages cannot
be cast to that type (C.virConnectPtr becomes libvirt.C.virConnectPtr)
- at least they could not as far as I could tell.

This function will solve both problems by being exported, and then
handling the type casting to C.virConnectPtr. This way someone can
just send us a pointer to the connection they created themselves,
and we can still use it in the library.