connection: remove dangerous SetPtr() and UnrefAndCloseConnection() methods
Those two changes break the API backward compatibility!

The first function was introduced in #81 to workaround the absence of
virConnectOpenAuth() method. This defeats the purpose of having an
opaque structure for a connection. Now that there is a mechanism to
handle callbacks, it should be easier to implement virConnectOpenAuth()
correctly.

The second function should not exist as there is no way to increment
reference of an existing connection from Go code (virConnectRef() is not
implemented). Therefore, if this function was needed, this should be
because the C code increments a reference to a connection but the
appropriate call to decrement it is absent. For example, we requested a
domain but we didn't free it. Therefore, the domain holds a reference to
the connection and the connection is not closed correctly. Forcing the
connection to close implies there is a leak somewhere (the domain should
be freed).