connection: move ancillary data to a separate struct
To implement error callbacks (in #96), the current error callback ID was
associated to the VirConnection structure on which the SetErrorFunc has
been called. However, if a user receives a connection from an event,
this connection won't have the right error callback ID.

Therefore, storing additional information in the VirConnection structure
seems a bit fragile. Either we should ensure that each connection is a
single VirConnection structure or we should store the data elsewhere.

This commit moves the data in a dedicated structure and a map is used to
associated a virConnectPtr to the appropriate ancillary data. Therefore,
for a connection, we have an unique C.virConnectPtr, zero or more
VirConnection struct and one virConnectionData struct.

The other solution (ensuring we have only one VirConnection struct for a
given connection) seemed far more complex to handle as it's hard to
avoid a copy by the user (we would have to force the user to only use
pointers).