connection: add RegisterCloseCallback() method
This enables to register one callback when the connection is closed
unexpectedly. Only one callback can be registered this way. This is
easier to handle this way.

It's quite hard to test if a connection is closed since it only works
for unexpected close (virConnectClose() won't do anything as the
callback will keep a reference to the connection). Therefore, we rely on
keepalive feature for that.