Add helpers for packing/unpacking typed parameters to/from structs
To automate the packing/unpacking of typed parameters to/from
structs, introduce a "typedParamsFieldInfo" struct. This is
used to provide pointers to struct fields that will provide/
receive the value of a typed parameter entry. Further helper
methods take a map of string->typedParamsFieldInfo to allow
automatic packing/unpacking of entire arrays of typed
parameters

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>