Fix handling of domain interface list data
The approach to handling the domain interface list data was broken and
only worked by luck when there was a single interface, with a single
address. If there were multiple interfaces or multiple addresses a
Go exception would result.

Rewrite it to use pointer arithmetic instead of trying to turn the C
array into a go slice

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>