Add NewConnectWithAuthDefault
It's currently not possible to use virConnectOpenAuth with
the convenience virConnectAuthPtrDefault callbacks. This can easily
be reimplemented in go if desired, but this commit adds a new
NewConnectWithAuthDefault method for convenience.

virConnectAuthPtrDefault is not wrapped in go as the implementation
would be quite awkward with the data being converted from C to go only
to be converted again to C.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>